Fix duck, develop Top page and add Link to navbar
