+++ src/ducks/top10.js
@@ -1,92 +1,41 @@
 import fetch from 'isomorphic-fetch'
 import Immutable from 'immutable'
 
-/* Flight Data */
-export const REQUEST_FLIGHT_DATA = 'REQUEST_FLIGHT_DATA'
-export const RECEIVE_FLIGHT_DATA = 'RECEIVE_FLIGHT_DATA'
-/* Positions */
-export const REQUEST_FLIGHT_POSITION = 'REQUEST_FLIGHT_POSITION'
-export const RECEIVE_FLIGHT_POSITION = 'RECEIVE_FLIGHT_POSITION'
+/* Top 10 */
+export const REQUEST_TOP10 = 'REQUEST_TOP10'
+export const RECEIVE_TOP10 = 'RECEIVE_TOP10'
 
-export function requestFlightData () {
+export function requestTop10 () {
   return {
-    type: REQUEST_FLIGHT_DATA
+    type: REQUEST_TOP10
   }
 }
 
-export function receiveFlightData (json) {
+export function receiveTop10 (json) {
   return {
-    type: RECEIVE_FLIGHT_DATA,
-    flightData: json
+    type: RECEIVE_TOP10,
+    top10: json
   }
 }
 
-export function fetchFlightData (flightId) {
+export function fetchTop10 () {
   return function (dispatch) {
-    dispatch(requestFlightData())
-    return fetch(`https://api.airstats.org/Flight/${flightId}`)
+    dispatch(requestTop10())
+    return fetch(`https://api.airstats.org/Statistics/Top10`)
       .then(response => response.json())
       .then(json =>
-        dispatch(receiveFlightData(Immutable.fromJS(json)))
+        dispatch(receiveTop10(Immutable.fromJS(json)))
       )
   }
 }
 
-export function requestFlightPosition () {
-  return {
-    type: REQUEST_FLIGHT_POSITION
-  }
-}
-
-export function receiveFlightPosition (json) {
-  return {
-    type: RECEIVE_FLIGHT_POSITION,
-    positions: json
-  }
-}
-
-export function fetchFlightPosition (flightId) {
-  return function (dispatch) {
-    dispatch(requestFlightPosition())
-    return fetch(`https://api.airstats.org/Flight/${flightId}/Positions`)
-      .then(response => response.json())
-      .then(json =>
-        dispatch(receiveFlightPosition(Immutable.fromJS(json)))
-    )
-  }
-}
-
-export default function flight (state = Immutable.fromJS({ isFetching: true, flight: { positions: [], flightData: {} } }), action) {
+export default function top10 (state = Immutable.fromJS({ isFetching: true, data: [] }), action) {
   switch (action.type) {
-    case REQUEST_FLIGHT_POSITION:
+    case REQUEST_TOP10:
       return state.setIn(['isFetching'], true)
-    case RECEIVE_FLIGHT_POSITION:
-      // Let's break this motherfucker down,
-      // First, we filter over the whole original list
-      // This will result in our stripped out list.
-      // Inside the predicate of the filter, take the value
-      // and go find the last index of the same lat/lon pairing. (equal lat/lons = no movement.)
-      // Once we find the last index, pass that value back to
-      // the filter, which will compare it against the value
-      // it's looking at's index. Filter goes from 0-2
-      // Find last goes from 2-0, any discrepency in the indexes when matched = duplicate, so we remove it.
-      let newPositions = action.positions.filter(
-        (value, index, iter) => {
-          return iter.findLastIndex(
-            (origValue) => {
-              return origValue.get('lat') === value.get('lat') && origValue.get('lon') === value.get('lon')
-            }
-          ) === index
-        }
-      )
+    case RECEIVE_TOP10:
       let newState = state.setIn(['isFetching'], false)
-      newState = newState.setIn(['flight', 'positions'], newPositions)
-      return newState
-    case REQUEST_FLIGHT_DATA:
-      return state.setIn(['isFetching'], true)
-    case RECEIVE_FLIGHT_DATA:
-      newState = state.setIn(['isFetching'], false)
-      newState = newState.setIn(['flight', 'flightData'], action.flightData)
+      newState = newState.set('data', action.top10)
       return newState
     default:
       return state
